Dense Matrix Unit Test Refactoring (#170)
Dense matrix unit test refactoring

* Enforcing specific matrix/vector types in implementation-specific unit tests
* Refactoring of testMatrix.cpp to use template testing function to quickly test different memory spaces
* Small changes to documentation and removal of commented out test
